+++ examples/dag-examples/dags/diamond.yaml
