+++ examples/dag-examples/networks/network3.yaml
