+++ examples/dag-examples/resources/cluster3.yaml
